Add Jenkins status icon and add plain branch name parameter
Removed the previous Travis status icon since it is no longer meaningful and
replaced it with an icond that shows the build status for master. The status
will always be for the master branch even if the README.md file is viewed on
GitHub on a different branch.

I also added another Jenkins build parameter that captures just the branch
name that is being built. Travis didn't have such an environment variable to
use.