updated format
@nicolevasilevsky 
I updated the format to make it clearer which steps belong together (e.g. preparing the release, deploying the release).
Please take a look at this new format, and if you like it, please merge. Otherwise, delete this PR.
Also: is the "initial setup" in the "prepare the release" part of "prepare the release" or part of "deploy the release"?