merged paroxysmal nocturnal hemoglobinuria
Fixes #4232

@nicolevasilevsky : I have a reasoning problem.
'paroxysmal nocturnal hemoglobinuria' is acquired.
'paroxysmal nocturnal hemoglobinuria 1' and 'paroxysmal nocturnal hemoglobinuria 2' are also acquired, but has also association to a gene (please see Ada's comment in the issue), which make them 'inherited'.
The issue in the reasoning is due to the fact that we have "inherited" and "acquired".
could you please review and advice? Thank you!