Add superclass to pregnancy loss, recurrent, 4
close #6867

@sabrina
I'm not totally certain about this. This superclass makes sense per Megan's suggestion (based on what is in the OMIM description) but it seems to be weird to have a disease that is a susceptibility and a disease. But it does seem to make sense in this case.

I wonder if maybe should remove the superclass 'fertility disorder'. Looks like I am the source and obviously, i am not an expert.