Add curated content resource property
Right now, we cannot agree community wide on a standard way to represent linkouts, not even whether or not they should be curated in the ontology at all. This is breaking a bit more with Mondo being a pure ontology vs a bit databasey, but right now, we just need a convenient way to link to curated content for our stakeholders.

Here, we add the property so we can start using it, even if we might change the ID to an OMO id in the future.

Context:
https://github.com/information-artifact-ontology/ontology-metadata/issues/165
https://github.com/monarch-initiative/mondo/issues/7071