#7640-Revision of Cornelia De Lange Syndrome (Real) (#7823)
* #7640-Revision of Cornelia De Lange Syndrome

MONDO:0016033, Cornelia De Lange Syndrome (CdLS)

Requested removing parent classes "Congenital Limb Malformation" and "Dysostosis".

According to PMID: 20301283, CdLS is characterized with "Limb abnormalities (either symmetric or asymmetric) ranging from severe reduction defects with complete absence of the forearms, to various forms of oligodactyly (missing digits) involving primarily the upper extremities, to small hands (micromelia) and phalangeal abnormalities (5th digit clinodactyly and short first metacarpal resulting in a proximally placed thumb) at the mild end". I have annotated this PMID to "Congenital Limb Malformation" and recommend it remains as a parent class.

Alternatively, CdLS appears to present as a Dysostoses by definition, however is not labeled that anywhere in my literature search. Dysostosis seems to be used to describe disorders with specific patterns of bone development. Considering the lack of usage of such term in the literature, as well as various limb malformation presentations seen in CdLS, I agree with the collaborator to remove Dysostosis as a parent class of CdLS.

* Updated source & excluded subclassof

Reviewed QC

- Changed dbxref to source
- Removed spaces between PMID: and following digits for added sources
- Added dysostosis as excluded subclassof