New term Dyneinopathy (#7894)
* New term Dyneinopathy

Closes #7652

Includes the following
- Definition
- 3 PMIDs for definition
- 2 Parent classes
- 2 Children classes

* Updated term dyneinopathy

- Removed first letter capitalization
- Added ORCID, Clingen, OMIM, and PMID as respective annotations

Did not yet address the addition of spinal muscular atrophy per the request of collaborator.

* Update MONDO ID & added annotation for dyneinopathy

Previous MONDO ID was 1040000, which has been used in other commits. Updated this ID to 1040031.

Also added the SubClassOf annotation: 'has material basis in germline mutation in' some DYNC1H1, along with sources for this.

* Update dyneinopathy definition, xref, and URL

Minor updates to the definition of dyneinopathy. Removed ncbi search URL and placed PMID of genereviews. Also created a term tracker item.