Refining subclass children, and other logical definitions
Removed split hand foot as a subclass, and made it a subclass of developmental defect of embryogenesis, as it can be this and its current subclass of disorder of development or morphogenesis. This made its subtype, split hand food type 3 a subclass, as it arises due to mutations specifically in TP63.

Also changed subclass of EEC syndrome to EEC type 3, as this one is specifically caused by mutations in TP63

Added PMIDs.

Fixed incorrect EquivalentTo definition in previous submission.