Update neurodevelopmental disorder with severe motor impairment and absent language
Closes #8265

According the literature, DHX-related complex neurodevelopmental disorder is present as neurodevelopmental disorder with severe motor impairment and absent language. This term has been updated to include this as an exact synonym, with a clingen preferred label.

Added github tracker, dbxref and sources to defs and subclasses and definition. Also added neurodevelopmental disorder with severe motor impairment and absent language as a subclass of complex neurodevelopmental disorder