Update porphyria due to ALA dehydratase deficiency (#8394)
* Update porphyria due to ALA dehydratase deficiency

Closes #8036

Updated the ClinGen preferred label to match the ClinGen IEM GCEPs current request. Updated definition, is essentially the same as the previous without typos or use of DALAD (submitter mentions this is an antiquated term). Included term tracker as well

* Remove rare labels for Keratoconus

Closes #8068

According to literature, keratoconus is not a rare disease (as high as 5% population has this disease). Removed rare labels. Added item tracker

* updates a few things as described in the comments

---------

Co-authored-by: Sabrina Toro <sabrina@tislab.org>