obsoleted MONDO:0012128
Closes #7841

Note: I added OMIM:608808 as a x-ref to MONDO:0012128, but without a source annotation (since it is not the omim term is not equivalent to MONDO:0019443 ).
I didn't like "MONDO:related" because OMIM is not "similar" to  MONDO:0019443. But I can add it if someone thinks it would be appropriate.