updates related GRIN2A
close #9063

I did not create a new term for GRIN2A-related disorders because investigation showed that only 2 out of 5 terms fit that criteria in an ontological standpoint.
‘epilepsy, focal, with speech disorder and with or without impaired intellectual development’ has been adjusted to be
a related synonym to 'Landau-Kleffner syndrome' but added as an
exact synonym to 'early-onset epileptic encephalopathy and intellectual disability due to GRIN2A mutation' based on Orphanet:289266. ‘continous spikes and waves during sleep’ should be merged with MONDO:0800501 'developmental and/or epileptic encephalopathy with spike-wave activation in sleep' because they represent the same concept based on Orphanet725 and OMIM:245570.