Update definition and synonyms for MONDO:0008030
Changes to reflect report in the Mondo paper review and having no support for a gene association to the FRG1 gene.

Revised the definition of facioscapulohumeral muscular dystrophy 1 to reflect association with D4Z4 contraction on chromosome 4q35. Updated synonym scopes and removed references to FRG1 mutation, aligning with current genetic understanding and external references.