Refresh imports (#9843)
* refresh imports

* add qc exclusion

* Update mondo-edit.obo

>MONDO:0011397 'autosomal dominant cerebellar ataxia, deafness and narcolepsy' --> relationship: disease_has_feature HP:0030050 {source="MONDO:Wikidata"} ! obsolete Narcolepsy

I removed the annotation because narcolepsy can be with or without cataplexy (the "replaced by" term for "obsolete Narcolepsy), and therefore it is unclear whether updating the annotation using the "replaced by" suggested term (cataplexy) would be correct without deeper investigation

>MONDO:0001485 'atypical depressive disorder' --> relationship: disease_has_feature HP:0100786 ! obsolete Hypersomnia

I removed this annotation since the annotation with the replacement term for "obsolete Hypersomnia" already exists

* update infectious agent in logical definitions

* 2terms with same logical definition

The NCBITaxon update lead to 2 terms having the same logical definition. I removed the logical for one of the terms. These terms will be merged in the context of https://github.com/monarch-initiative/mondo/issues/9850

---------

Co-authored-by: Sabrina Toro <sabrina@tislab.org>